+++ release_notes.md
@@ -1,3 +1,7 @@
+#### 1.1.0
+* added explicit download installer helper (DotnetDownloadInstaller)
+* add preview2 tooling shortcut (see preview2 sample)
+
 #### 1.0.0-rc2 - TBD
 * Initial version released to nuget
 * compatible with donet cli rc2